test_mock: remove conditional tests, move native coros
